Fix large alternation (7+ branches) failing to match full pattern
Closes #120. The root cause was UInt8 overflow in ASTNode.children_indexes.
Patterns with > 255 AST nodes (like the US NANPA area code pattern with
303 nodes) silently corrupted child indices, causing the parser to produce
a broken AST. The DFA compiler then generated a 1-byte match instead of
the full 10-byte phone number.

Changes:
- ast.mojo: Widen children_indexes from SIMD[uint8, 256] to
  SIMD[uint16, 512]. ChildrenIndexes type from List[UInt8] to List[UInt16].
  max_children from 256 to 512. Supports patterns with up to 65535 AST
  nodes and 512 children per node.
- parser.mojo: All child index variables changed from UInt8 to UInt16.
- optimizer.mojo: Added _has_nested_alternation and _group_contains_or
  to prevent classifying deeply nested alternation trees as SIMPLE
  (defense-in-depth against DFA compiler limitations).
- Regression test: full US NANPA pattern matching area codes 650, 212, 917.

+++ tests/test_ast.mojo
@@ -185,7 +185,7 @@ def test_GroupNode() raises:
     mut_regex.append_child(elem1)
     mut_regex.append_child(elem2)
 
-    var children_indexes = _make_children_indexes(UInt8(1), UInt8(2))
+    var children_indexes = _make_children_indexes(UInt16(1), UInt16(2))
     var group = GroupNode[MutAnyOrigin](
         regex,
         children_indexes=children_indexes,
@@ -208,7 +208,7 @@ def test_GroupNode_default_name() raises:
     # Add child to mutable regex
     mut_regex.append_child(elem)
 
-    var children_indexes = _make_children_indexes(UInt8(1))
+    var children_indexes = _make_children_indexes(UInt16(1))
     var group = GroupNode[MutAnyOrigin](
         regex,
         children_indexes=children_indexes,
